experiments/synthetic/requirements.txt
torch
torchvision
pandas
numpy
opencv-python
tqdm
requests
scikit-learn
wandb
ipykernel
PyYAML
